``` refactor(bench): drop _v2 suffix from bench modules
Rename the live v2 implementations to drop the _v2 suffix and remove
the dead v1 modules. Six module pairs collapse into single files:
coordinator, corpus, orchestrate, output, preflight, schemas. Tests
follow the same rename. Archive the v2 planning docs under
docs/archive/hippo-bench-v2-cleanup/.

```
