+++ brain/tests/test_bench_orchestrate.py
@@ -1,147 +1,121 @@
+"""Top-level orchestrator tests — dry-run manifest, output-dir creation,
+per-model failure isolation. Mocks `run_one_model` and the prod-pause RPC so
+the test doesn't need a live LM Studio or shadow brain.
+"""
+
+from __future__ import annotations
+
 import json
-import sqlite3
+from pathlib import Path
 from unittest.mock import patch
 
-from hippo_brain.bench.corpus import sample_from_hippo_db, write_corpus
+import pytest
+
+from hippo_brain.bench.coordinator import ModelRunResult
 from hippo_brain.bench.orchestrate import orchestrate_run
 
 
-def _seed_minimal_db(db_path):
-    conn = sqlite3.connect(db_path)
-    conn.executescript(
-        """
-        CREATE TABLE shell_events (
-            id INTEGER PRIMARY KEY, command TEXT, stdout TEXT, stderr TEXT,
-            duration_ms INTEGER, exit_code INTEGER, cwd TEXT, ts INTEGER
-        );
-        CREATE TABLE claude_sessions (
-            id INTEGER PRIMARY KEY, session_id TEXT, transcript TEXT,
-            message_count INTEGER, tool_calls_json TEXT, ts INTEGER
-        );
-        CREATE TABLE browser_events (
-            id INTEGER PRIMARY KEY, url TEXT, title TEXT, dwell_ms INTEGER,
-            scroll_depth REAL, ts INTEGER
-        );
-        CREATE TABLE workflow_runs (
-            id INTEGER PRIMARY KEY, repo TEXT, workflow_name TEXT,
-            conclusion TEXT, annotations_json TEXT, ts INTEGER
-        );
-        """
-    )
-    conn.execute(
-        "INSERT INTO shell_events (command, stdout, stderr, duration_ms, exit_code, cwd, ts)"
-        " VALUES (?, ?, ?, ?, ?, ?, ?)",
-        ("ls -la", "listing", "", 10, 0, "/tmp", 0),
-    )
-    conn.commit()
-    conn.close()
+@pytest.fixture
+def stub_corpus(tmp_path: Path) -> tuple[Path, Path]:
+    """Minimal corpus stubs — orchestrate reads the manifest with .get() fallbacks
+    so a present-but-bare manifest is enough for dry-run paths."""
+    sqlite = tmp_path / "corpus.sqlite"
+    sqlite.write_bytes(b"")  # bytes never read in dry-run / mocked paths
+    manifest = tmp_path / "corpus.manifest.json"
+    manifest.write_text(json.dumps({"corpus_content_hash": "sha256:test", "schema_version": 0}))
+    return sqlite, manifest
 
 
-def test_orchestrate_dry_run_produces_manifest_only(tmp_path):
-    fixture = tmp_path / "corpus-v1.jsonl"
-    manifest = tmp_path / "corpus-v1.manifest.json"
-    fixture.write_text("")
-    manifest.write_text('{"corpus_content_hash": "sha256:empty", "corpus_version": "corpus-v1"}')
+def test_dry_run_writes_manifest_then_run_end(stub_corpus, tmp_path):
+    """`--dry-run` short-circuits past preflight and per-model loop, producing
+    only a manifest record + a run_end record with reason='dry_run'."""
+    sqlite, manifest = stub_corpus
     out = tmp_path / "run.jsonl"
 
-    with patch("hippo_brain.bench.orchestrate.run_all_preflight") as mock_pf:
-        mock_pf.return_value = []
+    with patch("hippo_brain.bench.orchestrate.PauseRpcClient") as PauseClient:
+        PauseClient.return_value.probe_health.return_value = None
         result = orchestrate_run(
-            candidate_models=[],
-            corpus_version="corpus-v1",
-            fixture_path=fixture,
+            candidate_models=["m1"],
+            corpus_sqlite=sqlite,
             manifest_path=manifest,
-            base_url="http://localhost:1234/v1",
-            embedding_model="nomic",
             out_path=out,
-            timeout_sec=60,
-            self_consistency_events=0,
-            self_consistency_runs=0,
-            skip_checks=True,
             dry_run=True,
+            skip_prod_pause=True,
         )
+
     assert out.exists()
-    lines = out.read_text().splitlines()
-    records = [json.loads(line) for line in lines]
+    records = [json.loads(line) for line in out.read_text().splitlines() if line]
     assert records[0]["record_type"] == "run_manifest"
     assert records[-1]["record_type"] == "run_end"
-    assert records[-1]["reason"] in ("dry_run", "no_models")
+    assert records[-1]["reason"] == "dry_run"
     assert result.models_completed == []
+    assert result.preflight_aborted is False
 
 
-def test_orchestrate_creates_output_dir_before_preflight(tmp_path):
-    fixture = tmp_path / "corpus-v1.jsonl"
-    manifest = tmp_path / "corpus-v1.manifest.json"
-    fixture.write_text("")
-    manifest.write_text('{"corpus_content_hash": "sha256:empty", "corpus_version": "corpus-v1"}')
-    out = tmp_path / "nested" / "runs" / "run.jsonl"
+def test_orchestrate_creates_output_dir(stub_corpus, tmp_path):
+    """orchestrate_run mkdirs the JSONL parent before writing — even when the
+    operator points --out at a nested path that doesn't exist yet."""
+    sqlite, manifest = stub_corpus
+    out = tmp_path / "nested" / "subdir" / "run.jsonl"
 
-    with patch("hippo_brain.bench.orchestrate.run_all_preflight") as mock_pf:
-        mock_pf.return_value = []
+    with patch("hippo_brain.bench.orchestrate.PauseRpcClient") as PauseClient:
+        PauseClient.return_value.probe_health.return_value = None
         orchestrate_run(
             candidate_models=[],
-            corpus_version="corpus-v1",
-            fixture_path=fixture,
+            corpus_sqlite=sqlite,
             manifest_path=manifest,
-            base_url="http://localhost:1234/v1",
-            embedding_model="nomic",
             out_path=out,
-            timeout_sec=60,
-            self_consistency_events=0,
-            self_consistency_runs=0,
-            skip_checks=False,
             dry_run=True,
+            skip_prod_pause=True,
         )
 
-    assert out.parent.exists()
-    assert mock_pf.call_args.args[0] == out.parent
+    assert out.parent.is_dir()
+    assert out.exists()
 
 
-@patch("hippo_brain.bench.orchestrate.run_one_model")
-def test_orchestrate_isolates_failing_model(mock_run_one, tmp_path):
-    """When one model raises, later models still run; JSONL gets an error record."""
-    db = tmp_path / "hippo.db"
-    _seed_minimal_db(db)
-
-    fixture = tmp_path / "corpus-v1.jsonl"
-    manifest = tmp_path / "corpus-v1.manifest.json"
-    entries = sample_from_hippo_db(
-        db_path=db,
-        source_counts={"shell": 1, "claude": 0, "browser": 0, "workflow": 0},
-        seed=1,
-    )
-    write_corpus(entries, fixture, manifest, "corpus-v1", 1)
-
-    from hippo_brain.bench.coordinator import ModelRunResult
-
-    # First model raises; second returns a clean empty result.
-    mock_run_one.side_effect = [
-        RuntimeError("simulated lms crash"),
-        ModelRunResult(
-            model="m2",
-            attempts=[],
-            per_event_vectors=[],
-            peak_metrics={},
-            wall_clock_sec=1,
-            cooldown_timeout=False,
-        ),
-    ]
-
+def test_orchestrate_isolates_failing_model(stub_corpus, tmp_path):
+    """When run_one_model raises for one model, later models still run; the
+    JSONL gets a model_summary error record for the raiser. This is the BT-03/
+    BT-04 family contract — a per-model failure must not tank the rest of the
+    run.
+    """
+    sqlite, manifest = stub_corpus
     out = tmp_path / "run.jsonl"
-    result = orchestrate_run(
-        candidate_models=["m1", "m2"],
-        corpus_version="corpus-v1",
-        fixture_path=fixture,
-        manifest_path=manifest,
-        base_url="http://x/v1",
-        embedding_model="nomic",
-        out_path=out,
-        timeout_sec=5,
-        self_consistency_events=0,
-        self_consistency_runs=0,
-        skip_checks=True,
-        dry_run=False,
+
+    clean_result = ModelRunResult(
+        model="m2",
+        attempts=[],
+        per_event_vectors=[],
+        peak_metrics={},
+        wall_clock_sec=1,
+        cooldown_timeout=False,
+        process_ready_ms=10,
+        queue_drain_wall_clock_sec=0,
+        downstream_proxy={},
+        prod_brain_restarted_during_bench=False,
+        timeout_during_drain=False,
+        errors=[],
     )
+
+    with (
+        patch("hippo_brain.bench.orchestrate.run_one_model") as mock_run_one,
+        patch("hippo_brain.bench.orchestrate.PauseRpcClient") as PauseClient,
+    ):
+        PauseClient.return_value.probe_health.return_value = None
+        # First model raises; second returns cleanly.
+        mock_run_one.side_effect = [
+            RuntimeError("simulated lms crash"),
+            clean_result,
+        ]
+        result = orchestrate_run(
+            candidate_models=["m1", "m2"],
+            corpus_sqlite=sqlite,
+            manifest_path=manifest,
+            out_path=out,
+            skip_checks=True,
+            skip_prod_pause=True,
+            dry_run=False,
+        )
 
     assert result.models_completed == ["m2"]
     assert result.models_errored == ["m1"]
@@ -154,7 +128,29 @@ def test_orchestrate_isolates_failing_model(mock_run_one, tmp_path):
     assert m1_summary["attempts_total"] == 0
     assert any("simulated lms crash" in n for n in m1_summary["tier0_verdict"]["notes"])
 
-    # run_end record should be last
+    # run_end record should be last.
     assert records[-1]["record_type"] == "run_end"
     assert records[-1]["models_completed"] == ["m2"]
     assert records[-1]["models_errored"] == ["m1"]
+
+
+def test_orchestrate_no_models_emits_run_end_with_reason(stub_corpus, tmp_path):
+    """Empty candidate_models list short-circuits with reason='no_models'."""
+    sqlite, manifest = stub_corpus
+    out = tmp_path / "run.jsonl"
+
+    with patch("hippo_brain.bench.orchestrate.PauseRpcClient") as PauseClient:
+        PauseClient.return_value.probe_health.return_value = None
+        result = orchestrate_run(
+            candidate_models=[],
+            corpus_sqlite=sqlite,
+            manifest_path=manifest,
+            out_path=out,
+            skip_checks=True,
+            skip_prod_pause=True,
+            dry_run=False,
+        )
+
+    records = [json.loads(line) for line in out.read_text().splitlines() if line]
+    assert records[-1]["reason"] == "no_models"
+    assert result.models_completed == []
