+++ brain/tests/test_bench_runner.py
@@ -2,35 +2,7 @@ from unittest.mock import patch
 
 from hippo_brain.bench.corpus import CorpusEntry
 from hippo_brain.bench.enrich_call import CallResult
-from hippo_brain.bench.runner import run_model_main_pass, run_self_consistency_pass
-
-
-@patch("hippo_brain.bench.runner.call_enrichment")
-def test_main_pass_produces_one_attempt_per_event(mock_call):
-    mock_call.return_value = CallResult(
-        raw_output=(
-            '{"summary": "ok", "intent": "x", "outcome": "success", '
-            '"entities": {"projects": [], "tools": [], "files": [], "services": [], "errors": []}}'
-        ),
-        ttft_ms=None,
-        total_ms=100,
-        timeout=False,
-    )
-    entries = [
-        CorpusEntry(event_id="e1", source="shell", redacted_content="ls"),
-        CorpusEntry(event_id="e2", source="shell", redacted_content="pwd"),
-    ]
-    attempts = run_model_main_pass(
-        base_url="http://x",
-        model="m1",
-        entries=entries,
-        timeout_sec=10,
-        metrics_snapshot=lambda: {"lmstudio_rss_mb": 100.0},
-        temperature=0.7,
-    )
-    assert len(attempts) == 2
-    assert all(a.purpose == "main" for a in attempts)
-    assert attempts[0].event["event_id"] == "e1"
+from hippo_brain.bench.runner import run_self_consistency_pass
 
 
 @patch("hippo_brain.bench.runner.call_embedding")
@@ -60,38 +32,3 @@ def test_self_consistency_pass_embeds_each_output(mock_call, mock_embed):
     assert len(attempts) == 2 * 3
     assert len(per_event_vectors) == 2
     assert all(len(v) == 3 for v in per_event_vectors)
-
-
-@patch("hippo_brain.bench.runner.call_enrichment")
-def test_main_pass_marks_timeouts(mock_call):
-    mock_call.return_value = CallResult(raw_output="", ttft_ms=None, total_ms=1000, timeout=True)
-    entries = [CorpusEntry(event_id="e1", source="shell", redacted_content="ls")]
-    attempts = run_model_main_pass(
-        base_url="http://x",
-        model="m1",
-        entries=entries,
-        timeout_sec=1,
-        metrics_snapshot=lambda: {},
-        temperature=0.7,
-    )
-    assert attempts[0].timeout is True
-
-
-@patch("hippo_brain.bench.runner.time.monotonic_ns", return_value=123456789)
-@patch("hippo_brain.bench.runner.call_enrichment")
-def test_main_pass_captures_start_timestamps_before_call(mock_call, mock_monotonic):
-    def fake_call(**_kwargs):
-        return CallResult(raw_output="", ttft_ms=None, total_ms=10, timeout=True)
-
-    mock_call.side_effect = fake_call
-    entries = [CorpusEntry(event_id="e1", source="shell", redacted_content="ls")]
-    attempts = run_model_main_pass(
-        base_url="http://x",
-        model="m1",
-        entries=entries,
-        timeout_sec=1,
-        metrics_snapshot=lambda: {},
-        temperature=0.7,
-    )
-    assert attempts[0].timestamps["start_iso"]
-    assert attempts[0].timestamps["start_monotonic_ns"] == 123456789
