+++ docs/archive/hippo-bench-v2-cleanup/2026-04-21-hippo-bench-mvp.md
