+++ docs/archive/hippo-bench-v2-cleanup/2026-04-27-hippo-bench-v2-implementation.md
